modificacion username en views

+++ src/componentes/DashboardNavbar.jsx
@@ -16,8 +16,12 @@ import {
   PlusCircle,
   Clipboard
 } from "lucide-react";
+import { API_BASE_URL } from "../config";
+import { useNavigate } from "react-router-dom";
 
 const DashboardNavbar = () => {
+  const navigate = useNavigate();
+  const [busqueda, setBusqueda] = useState("");
   const [menuPerfilAbierto, setMenuPerfilAbierto] = useState(false);
   const [menuLateralAbierto, setMenuLateralAbierto] = useState(false);
   const [busquedaActiva, setBusquedaActiva] = useState(false);
@@ -46,9 +50,40 @@ const DashboardNavbar = () => {
     setMenuLateralAbierto(!menuLateralAbierto);
   };
 
-  const cerrarSesion = () => {
-    localStorage.clear();
-    window.location.href = '/';
+  const manejarBusqueda = (e) => {
+    if (e.key === "Enter" && busqueda.trim() !== "") {
+      navigate(`/resultados-busqueda?q=${encodeURIComponent(busqueda.trim())}`);
+    }
+  };
+
+  const handleLogout = async () => {
+    try {
+      const tokenSesion = localStorage.getItem("token_sesion");
+  
+      if (!tokenSesion) {
+        console.error("No se encontró el token de sesión en localStorage.");
+        return;
+      }
+  
+      const response = await fetch(`${API_BASE_URL}/logout/`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ token_sesion: tokenSesion }),
+      });
+  
+      const data = await response.json();
+  
+      if (response.ok) {
+        localStorage.clear();
+        window.location.href = '/login';
+      } else {
+        console.error('Error al cerrar sesión:', data.error);
+      }
+    } catch (error) {
+      console.error('Error de red:', error);
+    }
   };
 
   return (
@@ -95,6 +130,9 @@ const DashboardNavbar = () => {
               <input
                 type="text"
                 placeholder="Buscar o ir a..."
+                value={busqueda}
+                onChange={(e) => setBusqueda(e.target.value)}
+                onKeyDown={manejarBusqueda}
                 className="bg-transparent outline-none text-md text-white placeholder-white/60 w-full"
               />
               <kbd className="ml-2 text-white/50 text-xs border border-white/30 rounded px-1">/</kbd>
@@ -226,7 +264,7 @@ const DashboardNavbar = () => {
             <ul className="space-y-3 text-red-600 text-sm font-medium">
               <li
                 className="flex items-center gap-2 hover:text-red-700 transition cursor-pointer"
-                onClick={cerrarSesion}
+                onClick={handleLogout}
               >
                 <LogOut size={18} /> Cerrar sesión
               </li>
